add jwt token and authorization

+++ backend/routes/product.js
@@ -1,11 +1,19 @@
 const Product = require("../models/Product");
 
 const router = require("express").Router();
+const {
+  verifyToken,
+  verifyTokenAndAuthorizationEntity,
+  verifyTokenAndAdmin,
+} = require("./verifyToken");
 
 //CREATE
 //verificata
-router.post("/add-product", async (req, res) => {
-  const newProduct = new Product(req.body);
+router.post("/add-product", verifyToken, async (req, res) => {
+  const newProduct = new Product({
+    ...req.body,
+    userId: req.user.id,
+  });
 
   try {
     const savedProduct = await newProduct.save();
@@ -17,46 +25,58 @@ router.post("/add-product", async (req, res) => {
 
 //UPDATE
 //verificata
-router.put("/:id", async (req, res) => {
-  try {
-    const updatedProduct = await Product.findByIdAndUpdate(
-      req.params.id,
-      {
-        $set: req.body,
-      },
-      { new: true }
-    );
-    res.status(200).json(updatedProduct);
-  } catch (err) {
-    res.status(500).json(err);
+router.put(
+  "/:id",
+  verifyTokenAndAuthorizationEntity("Product"),
+  async (req, res) => {
+    try {
+      const updatedProduct = await Product.findByIdAndUpdate(
+        req.params.id,
+        {
+          $set: req.body,
+        },
+        { new: true }
+      );
+      res.status(200).json(updatedProduct);
+    } catch (err) {
+      res.status(500).json(err);
+    }
   }
-});
+);
 
 //DELETE
 //verificata
-router.delete("/:id", async (req, res) => {
-  try {
-    await Product.findByIdAndDelete(req.params.id);
-    res.status(200).json("Product has been deleted...");
-  } catch (err) {
-    res.status(500).json(err);
+router.delete(
+  "/:id",
+  verifyTokenAndAuthorizationEntity("Product"),
+  async (req, res) => {
+    try {
+      await Product.findByIdAndDelete(req.params.id);
+      res.status(200).json("Product has been deleted...");
+    } catch (err) {
+      res.status(500).json(err);
+    }
   }
-});
+);
 
 //GET PRODUCT
 //verificata
-router.get("/:id", async (req, res) => {
-  try {
-    const product = await Product.findById(req.params.id);
-    res.status(200).json(product);
-  } catch (err) {
-    res.status(500).json(err);
+router.get(
+  "/:id",
+  verifyTokenAndAuthorizationEntity("Product"),
+  async (req, res) => {
+    try {
+      const product = await Product.findById(req.params.id);
+      res.status(200).json(product);
+    } catch (err) {
+      res.status(500).json(err);
+    }
   }
-});
+);
 
 //GET ALL PRODUCTS
 //verificata
-router.get("/", async (req, res) => {
+router.get("/", verifyTokenAndAdmin, async (req, res) => {
   const qNew = req.query.new;
   const qCategory = req.query.category;
   try {
